+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 **Internal**
 
 - Allow deferred lengths to the `/upload` endpoint when the sender is trusted. ([#5658](https://github.com/getsentry/relay/pull/5658))
+- Use new processor architecture to process client reports. ([#5686](https://github.com/getsentry/relay/pull/5686))
 
 ## 26.2.1